Separated Telescope installation into its own command.
Made adjustments to support future independent dependencies installations
